+++ admin/app/controllers/settings_controller.ts
@@ -41,6 +41,10 @@ export default class SettingsController {
     return inertia.render('settings/caches', { caches })
   }
 
+  async translation({ inertia }: HttpContext) {
+    return inertia.render('settings/translation')
+  }
+
   async legal({ inertia }: HttpContext) {
     return inertia.render('settings/legal')
   }
@@ -120,9 +124,9 @@ export default class SettingsController {
   }
 
   async getSetting({ request, response }: HttpContext) {
-    const { key } = await getSettingSchema.validate({ key: request.qs().key });
-    const value = await KVStore.getValue(key);
-    return response.status(200).send({ key, value });
+    const { key } = await getSettingSchema.validate({ key: request.qs().key })
+    const value = await KVStore.getValue(key)
+    return response.status(200).send({ key, value })
   }
 
   async updateSetting({ request, response }: HttpContext) {